delete and assign/unassign tags
