+++ app/cmd/routes.go
@@ -117,8 +117,8 @@ func routes(r *web.Engine) *web.Engine {
 			private.Delete("/api/v1/posts/:number/support", apiv1.RemoveSupporter())
 			private.Post("/api/v1/posts/:number/subscription", apiv1.Subscribe())
 			private.Delete("/api/v1/posts/:number/subscription", apiv1.Unsubscribe())
-			private.Post("/api/posts/:number/tags/:slug", handlers.AssignTag())
-			private.Delete("/api/posts/:number/tags/:slug", handlers.UnassignTag())
+			private.Post("/api/v1/posts/:number/tags/:slug", apiv1.AssignTag())
+			private.Delete("/api/v1/posts/:number/tags/:slug", apiv1.UnassignTag())
 			private.Delete("/api/user", handlers.DeleteUser())
 			private.Post("/api/user/settings", handlers.UpdateUserSettings())
 			private.Post("/api/user/change-email", handlers.ChangeUserEmail())
@@ -142,7 +142,7 @@ func routes(r *web.Engine) *web.Engine {
 
 			private.Get("/admin/export", handlers.Page("Export · Site Settings", ""))
 			private.Get("/admin/export/posts.csv", handlers.ExportPostsToCSV())
-			private.Delete("/api/posts/:number", handlers.DeletePost())
+			private.Delete("/api/v1/posts/:number", apiv1.DeletePost())
 			private.Post("/api/admin/settings/general", handlers.UpdateSettings())
 			private.Post("/api/admin/settings/advanced", handlers.UpdateAdvancedSettings())
 			private.Post("/api/admin/settings/privacy", handlers.UpdatePrivacy())
