+++ app/handlers/apiv1/post.go
@@ -93,6 +93,23 @@ func SetResponse() web.HandlerFunc {
 	}
 }
 
+// DeletePost deletes an existing post of current tenant
+func DeletePost() web.HandlerFunc {
+	return func(c web.Context) error {
+		input := new(actions.DeletePost)
+		if result := c.BindTo(input); !result.Ok {
+			return c.HandleValidation(result)
+		}
+
+		err := c.Services().Posts.SetResponse(input.Post, input.Model.Text, models.PostDeleted)
+		if err != nil {
+			return c.Failure(err)
+		}
+
+		return c.Ok(web.Map{})
+	}
+}
+
 // AddSupporter adds current user to given post list of supporters
 func AddSupporter() web.HandlerFunc {
 	return func(c web.Context) error {
