+++ app/handlers/tag_test.go
@@ -9,7 +9,6 @@ import (
 	. "github.com/getfider/fider/app/pkg/assert"
 	"github.com/getfider/fider/app/pkg/errors"
 	"github.com/getfider/fider/app/pkg/mock"
-	"github.com/getfider/fider/app/pkg/web"
 )
 
 func TestCreateTagHandler_ValidRequests(t *testing.T) {
@@ -169,86 +168,4 @@ func TestDeleteExistingTagHandler_Collaborator(t *testing.T) {
 	Expect(tag).IsNotNil()
 	Expect(err).IsNil()
 
-}
-
-func TestAssignTagHandler_Success(t *testing.T) {
-	RegisterT(t)
-
-	server, services := mock.NewServer()
-	services.SetCurrentTenant(mock.DemoTenant)
-	services.SetCurrentUser(mock.JonSnow)
-	tag, _ := services.Tags.Add("Bug", "0000FF", true)
-	post, _ := services.Posts.Add("Post Title", "Post Description")
-
-	status, _ := server.
-		AsUser(mock.JonSnow).
-		AddParam("slug", tag.Slug).
-		AddParam("number", post.Number).
-		Execute(handlers.AssignTag())
-
-	tags, err := services.Tags.GetAssigned(post)
-	Expect(status).Equals(http.StatusOK)
-	Expect(err).IsNil()
-	Expect(tags[0]).Equals(tag)
-}
-
-func TestAssignTagHandler_UnknownTag(t *testing.T) {
-	RegisterT(t)
-
-	server, _ := mock.NewServer()
-
-	status, _ := server.
-		AsUser(mock.JonSnow).
-		AddParam("slug", "bug").
-		AddParam("number", 1).
-		Execute(handlers.AssignTag())
-
-	Expect(status).Equals(http.StatusNotFound)
-}
-
-func TestAssignOrUnassignTagHandler_Unauthorized(t *testing.T) {
-	RegisterT(t)
-
-	var testCases = []web.HandlerFunc{
-		handlers.AssignTag(),
-		handlers.UnassignTag(),
-	}
-
-	for _, handler := range testCases {
-		server, services := mock.NewServer()
-		services.SetCurrentTenant(mock.DemoTenant)
-		services.SetCurrentUser(mock.JonSnow)
-		tag, _ := services.Tags.Add("Bug", "0000FF", true)
-		post, _ := services.Posts.Add("Post Title", "Post Description")
-
-		status, _ := server.
-			AsUser(mock.AryaStark).
-			AddParam("slug", tag.Slug).
-			AddParam("number", post.Number).
-			Execute(handler)
-
-		Expect(status).Equals(http.StatusForbidden)
-	}
-}
-
-func TestUnassignTagHandler_Success(t *testing.T) {
-	RegisterT(t)
-
-	server, services := mock.NewServer()
-	services.SetCurrentTenant(mock.DemoTenant)
-	services.SetCurrentUser(mock.JonSnow)
-	tag, _ := services.Tags.Add("Bug", "0000FF", true)
-	post, _ := services.Posts.Add("Post Title", "Post Description")
-	services.Tags.AssignTag(tag, post)
-
-	status, _ := server.
-		AsUser(mock.JonSnow).
-		AddParam("slug", tag.Slug).
-		AddParam("number", post.Number).
-		Execute(handlers.UnassignTag())
-
-	tags, err := services.Tags.GetAssigned(post)
-	Expect(status).Equals(http.StatusOK)
-	Expect(err).IsNil()
-	Expect(tags).HasLen(0)
 }
